requirements.txt
flask~=3.1.2
opencv-python~=4.13.0.90
pillow~=12.1.0
Werkzeug
setuptools
picamera2
numpy~=2.4.1
piexif~=1.1.3
cv2_enumerate_cameras~=1.3.3